requirements.txt
# 파이썬
python-dotenv==1.0.1

# 의존성 관련
pip-tools==7.4.1
pipreqs==0.5.0

# Flask
Flask==3.1.0

# 테스팅
pytest==8.3.4
pytest-mock==3.14.0


# 웹 크롤링
requests==2.32.3
beautifulsoup4==4.12.3

